fix: correggi 3 errori di compilazione del kernel
- interrupts.rs: indice IDT cambiato da usize a u8 (E0277)
- memory/mod.rs: cast heap_start a *mut u8 per init() (E0308)
- panic.rs: rimosso .and_then() obsoleto su PanicMessage (E0599)
- gdt.rs: rimosso import SegmentSelector inutilizzato
- ui/mod.rs: rimossi import Renderer e InputEvent inutilizzati
- desktop.rs: prefisso _ su variabili h ed ext inutilizzate

+++ kernel/src/interrupts.rs
@@ -37,7 +37,7 @@ lazy_static! {
 
         idt.page_fault.set_handler_fn(page_fault_handler);
 
-        idt[InterruptIndex::Keyboard as usize]
+        idt[InterruptIndex::Keyboard as u8]
             .set_handler_fn(keyboard_interrupt_handler);
 
         idt
